moved `naturalCompare` function to new `AnimeTokiUtils.kt` file for better maintanence
