+++ src/en/animetoki/src/eu/kanade/tachiyomi/animeextension/en/animetoki/extractors/CloudExtractor.kt
@@ -3,6 +3,7 @@ package eu.kanade.tachiyomi.animeextension.en.animetoki.extractors
 import android.util.Base64
 import android.util.Log
 import eu.kanade.tachiyomi.animeextension.en.animetoki.CloudFileResponse
+import eu.kanade.tachiyomi.animeextension.en.animetoki.naturalCompare
 import eu.kanade.tachiyomi.animesource.model.SEpisode
 import eu.kanade.tachiyomi.network.POST
 import kotlinx.serialization.decodeFromString
@@ -130,34 +131,5 @@ class CloudExtractor(private val client: OkHttpClient, private val headers: Head
             Log.e("AnimeToki", "Error parsing cloud folder: $folderUrl", e)
             Log.e("AnimeToki", "Response Body: $responseBody")
         }
-    }
-
-    private fun naturalCompare(a: String, b: String): Int {
-        var ia = 0
-        var ib = 0
-        while (ia < a.length && ib < b.length) {
-            val charA = a[ia]
-            val charB = b[ib]
-            if (charA.isDigit() && charB.isDigit()) {
-                val startA = ia
-                while (ia < a.length && a[ia].isDigit()) ia++
-                val startB = ib
-                while (ib < b.length && b[ib].isDigit()) ib++
-
-                val numA = a.substring(startA, ia).toLongOrNull() ?: 0L
-                val numB = b.substring(startB, ib).toLongOrNull() ?: 0L
-                val cmp = numA.compareTo(numB)
-                if (cmp != 0) return cmp
-
-                val lenCmp = (ia - startA).compareTo(ib - startB)
-                if (lenCmp != 0) return lenCmp
-            } else {
-                val cmp = charA.lowercaseChar().compareTo(charB.lowercaseChar())
-                if (cmp != 0) return cmp
-                ia++
-                ib++
-            }
-        }
-        return a.length.compareTo(b.length)
     }
 }
